backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy>=2.0.30,<3.0
alembic==1.12.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Logging and monitoring
structlog==23.2.0
rich==13.7.0

# Configuration
python-dotenv==1.0.0

# Validation
pydantic==2.5.0
email-validator==2.1.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Additional utilities
aiofiles==23.2.1
jinja2==3.1.2
